Item 9 Cont. - Students can now join a club.
Continued work on 9. Student-proposed functionality #1 - Club stuff. A student can now join a club with they input their correct password, the given club actually exists, and if the student is not already in the club.

I also added missing mysqli_close($myconnection); lines to the bottom of item4 and item9create (renamed from item9). I also added the HTML form for leaving a club, but it doesn't do anything yet.